refactor(core): split built-in tools
